Rebase + fixes
- Rebased on rc3 main
- Removed debug prints
- Added tqdm to reqs
- Added more error handling for stats to csv (region.axis_minor_length value errors)

+++ napari_cellseg3d/interface.py
@@ -496,10 +496,10 @@ class DoubleIncrementCounter(QDoubleSpinBox):
         fixed: Optional[bool] = True,
     ):
         """Args:
-        min (Optional[int]): minimum value, defaults to 0
-        max (Optional[int]): maximum value, defaults to 10
-        default (Optional[int]): default value, defaults to 0
-        step (Optional[int]): step value, defaults to 1
+        min (Optional[float]): minimum value, defaults to 0
+        max (Optional[float]): maximum value, defaults to 10
+        default (Optional[float]): default value, defaults to 0
+        step (Optional[float]): step value, defaults to 1
         parent: parent widget, defaults to None
         fixed (bool): if True, sets the QSizePolicy of the spinbox to Fixed"""
